Only deploy .my.cnf file on galera servers
The original bug was opened because .my.cnf (with sensitive root credentials)
was being deployed on compute hosts without needing to be there. This patch
restricts the deployment of .my.cnf files to the galera server containers
only. This improves security while still allowing deployers to diagnose galera
server issues on the galera server containers themselves.

Closes-bug: 1510596

Change-Id: I42f39cbfcb02b97846894592d642ac0c58a82b02